fixed some gc issues. unit tests still failing?

+++ src/main/sm_global.c
@@ -64,7 +64,7 @@ sm_stack *sm_global_lex_stack(sm_stack *replacement) {
 
 // Return the name of this type.
 // Must be synchronized with enum sm_object_type
-char *sm_global_type_name(uint16_t which) {
+char *sm_global_type_name(uint32_t which) {
   static char *response[] = {"float", "expr", "prim",   "str",   "sym",   "cx",  "node",
                              "ptr",   "meta", "space",  "fun",   "param", "lcl", "l",
                              "err",   "self", "return", "block", "stack", "?"};
@@ -73,14 +73,14 @@ char *sm_global_type_name(uint16_t which) {
 
 // Return the length of the name of this type.
 // Must be synchronized with enum sm_object_type
-uint16_t sm_global_type_name_len(uint16_t which) {
-  static uint16_t response_len[] = {5, 4, 4, 3, 3, 2, 4, 3, 4, 5, 3, 5, 3, 1, 3, 4, 6, 5, 5, 1};
+uint32_t sm_global_type_name_len(uint32_t which) {
+  static uint32_t response_len[] = {5, 4, 4, 3, 3, 2, 4, 3, 4, 5, 3, 5, 3, 1, 3, 4, 6, 5, 5, 1};
   return response_len[which];
 }
 
 // Primitive_names
-char *sm_global_fn_name(uint16_t which) {
-  const uint16_t num_functions = sm_global_num_fns();
+char *sm_global_fn_name(uint32_t which) {
+  const uint32_t num_functions = sm_global_num_fns();
   // Should be syncronized with enum SM_EXPR_TYPE
   static char *response[] = {
     "exit",         // SM_EXIT_EXPR
@@ -272,7 +272,7 @@ char *sm_global_fn_name(uint16_t which) {
 }
 
 // Corresponding string length of the string that would come from the sm_global_fn_name(which)
-uint32_t sm_global_fn_name_len(uint16_t which) {
+uint32_t sm_global_fn_name_len(uint32_t which) {
   static uint64_t response_len[] = {
     4, 4,  5, 4,  2, 2,  3,  3, 1, 1, 1, 1, 1,  2, 1, 1, 1, 1,  1, 1, 3, 3, 3, 4, 4, 4,
     4, 4,  4, 5,  5, 5,  3,  3, 3, 4, 4, 4, 4,  4, 4, 5, 5, 5,  2, 3, 3, 4, 3, 4, 4, 3,
@@ -287,7 +287,7 @@ uint32_t sm_global_fn_name_len(uint16_t which) {
 }
 
 uint32_t sm_global_num_fns() {
-  static const uint16_t num_fns = 176;
+  static const uint32_t num_fns = 176;
   return num_fns;
 }
